feat: move sidebar components
